Backwards-compatible rename Operations::Form to Operations::Form::Base
